Shared Prefrence: On first launch, the app shows the Splash screen, then the Start screen.
On the next launches, it shows the Splash screen, then goes directly to Login Screen.

If the user is not registered, the app navigates to the Register screen.

If the user is already registered, the app navigates to Home Screen.

Language Support:

The app supports both Arabic and English.

The selected language is saved using Shared Preferences.

The app loads the saved language on startup.

The interface updates immediately when the language is changed.

Cubit and Shared Data:

The app uses Cubit to manage shared data like the user name.

AddTask screen:

When a task is added from the Add Task screen, it appears in the Home screen.

Image Picker Manager:

The app uses an image picker to select images from the device.